+++ src/main/java/kitchenpos/order/dto/OrderTableCreateRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.dto;
+package kitchenpos.order.dto;
 
 public class OrderTableCreateRequest {
     private int numberOfGuests;
