+++ src/main/java/kitchenpos/order/dto/TableGuestChangeRequest.java
@@ -1,4 +1,4 @@
-package kitchenpos.table.dto;
+package kitchenpos.order.dto;
 
 public class TableGuestChangeRequest {
     private final int numberOfGuests;
