+++ src/main/java/kitchenpos/order/ui/TableRestController.java
@@ -1,12 +1,12 @@
-package kitchenpos.table.ui;
+package kitchenpos.order.ui;
 
 import java.net.URI;
 import java.util.List;
-import kitchenpos.table.application.TableService;
-import kitchenpos.table.dto.OrderTableChangeEmptyRequest;
-import kitchenpos.table.dto.OrderTableCreateRequest;
-import kitchenpos.table.dto.OrderTableResponse;
-import kitchenpos.table.dto.TableGuestChangeRequest;
+import kitchenpos.order.application.TableService;
+import kitchenpos.order.dto.OrderTableChangeEmptyRequest;
+import kitchenpos.order.dto.OrderTableCreateRequest;
+import kitchenpos.order.dto.OrderTableResponse;
+import kitchenpos.order.dto.TableGuestChangeRequest;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
