+++ src/main/java/kitchenpos/tablegroup/domain/Tables.java
@@ -1,8 +1,9 @@
-package kitchenpos.table.domain;
+package kitchenpos.tablegroup.domain;
 
 import java.util.List;
 import java.util.Objects;
 import java.util.stream.Collectors;
+import kitchenpos.order.domain.OrderTable;
 
 public class Tables {
     private final List<OrderTable> value;
