+++ src/main/java/kitchenpos/tablegroup/dto/TableGroupResponse.java
@@ -1,8 +1,9 @@
-package kitchenpos.table.dto;
+package kitchenpos.tablegroup.dto;
 
 import java.time.LocalDateTime;
 import java.util.List;
-import kitchenpos.table.domain.TableGroup;
+import kitchenpos.order.dto.OrderTableResponse;
+import kitchenpos.tablegroup.domain.TableGroup;
 
 public class TableGroupResponse {
     private final Long id;
