+++ src/main/java/kitchenpos/tablegroup/repository/TableGroupRepository.java
@@ -1,7 +1,7 @@
-package kitchenpos.table.repository;
+package kitchenpos.tablegroup.repository;
 
 import java.util.List;
-import kitchenpos.table.domain.TableGroup;
+import kitchenpos.tablegroup.domain.TableGroup;
 
 public interface TableGroupRepository {
     TableGroup save(TableGroup entity);
