+++ src/test/java/kitchenpos/application/ServiceTest.java
@@ -7,14 +7,14 @@ import kitchenpos.menu.application.MenuGroupService;
 import kitchenpos.menu.application.MenuService;
 import kitchenpos.order.application.OrderService;
 import kitchenpos.product.application.ProductService;
-import kitchenpos.table.application.TableGroupService;
-import kitchenpos.table.application.TableService;
+import kitchenpos.tablegroup.application.TableGroupService;
+import kitchenpos.order.application.TableService;
 import kitchenpos.menu.repository.MenuRepository;
 import kitchenpos.menu.repository.MenuGroupRepository;
 import kitchenpos.menu.repository.MenuProductRepository;
-import kitchenpos.table.repository.OrderTableRepository;
+import kitchenpos.order.repository.OrderTableRepository;
 import kitchenpos.product.repository.ProductRepository;
-import kitchenpos.table.repository.TableGroupRepository;
+import kitchenpos.tablegroup.repository.TableGroupRepository;
 import org.junit.jupiter.api.BeforeEach;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
